+++ BitURL/FlashNotificationViewController.swift
@@ -1,0 +1,19 @@
+//
+//  FlashNotificationViewController.swift
+//  BitURL
+//
+//  Created by Roberto Mauro on 01/11/2019.
+//  Copyright © 2019 Roberto Mauro. All rights reserved.
+//
+
+import Cocoa
+
+class FlashNotificationViewController: NSViewController {
+
+    @IBOutlet var notificationTitleTextField: NSTextField?
+    @IBOutlet var notificationIconView: NSImageView?
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}